+++ src/commonTest/kotlin/kotools/types/experimental/NonZeroIntTest.kt
@@ -10,9 +10,10 @@ import kotools.types.number.NonZeroInt
 import kotools.types.number.toNonZeroInt
 import kotlin.test.Test
 import kotlin.test.assertEquals
+import kotlin.test.assertTrue
 
+@ExperimentalKotoolsTypesApi
 class NonZeroIntTest {
-    @ExperimentalKotoolsTypesApi
     @Test
     fun unaryMinus_should_pass() {
         val number: NonZeroInt = NonZeroInt.random()
@@ -20,6 +21,24 @@ class NonZeroIntTest {
         val expected: NonZeroInt = number.toInt()
             .unaryMinus()
             .toNonZeroIntOrFailure()
+        assertEquals(expected, actual)
+    }
+
+    @Test
+    fun positiveRange_should_go_from_1_included_to_Int_MAX_VALUE_included() {
+        val actual: NotEmptyRange<NonZeroInt> = NonZeroInt.positiveRange
+        assertTrue { actual.start is InclusiveBound }
+        val expectedStartValue: NonZeroInt = 1.toNonZeroIntOrFailure()
+        assertEquals(expectedStartValue, actual.start.value)
+        assertTrue { actual.end is InclusiveBound }
+        val expectedEndValue: NonZeroInt = Int.MAX_VALUE.toNonZeroIntOrFailure()
+        assertEquals(expectedEndValue, actual.end.value)
+    }
+
+    @Test
+    fun positiveRange_documentation_should_be_valid() {
+        val actual: String = NonZeroInt.positiveRange.toString()
+        val expected = "[1;2147483647]"
         assertEquals(expected, actual)
     }
 }