+++ CHANGELOG.md
@@ -70,7 +70,8 @@ All notable changes to this project will be documented in this file.
     `NotBlankString`
   - `Char.plus(NotBlankString)` operation
   - `range` property for companion object of `PositiveInt`, `NegativeInt`,
-    `StrictlyPositiveInt` and `StrictlyNegativeInt`.
+    `StrictlyPositiveInt` and `StrictlyNegativeInt`
+  - `NonZeroInt.Companion.positiveRange` property.
 - Opt-in message of experimental annotations (pull request
   [#328](https://github.com/kotools/types/pull/328)).
 - Documentation of types in API reference and in README (commit